+++ apps/app/src/components/Sidebar/PageCreateButton/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './use-create-new-page';
+export * from './use-create-todays-memo';
